Move ProgressListener and ConsoleListener implementations to dedicated classes
